audio-processing/requirements.txt
accelerate==0.26.1
datasets==2.18.0
evaluate==0.4.1
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
PyYAML==6.0.1
scikit-learn==1.4.0
spacy==3.7.2
tokenizers==0.19.1
torch==2.2.0
tqdm==4.66.4
transformers==4.40.0

# Pipeline dependencies
openai-whisper>=20231117
ffmpeg-python>=0.2.0
librosa>=0.10.0
soundfile>=0.12.1
pyaudio>=0.2.11

# Web service dependencies
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio>=5.8.0
requests>=2.31.0

# Additional utilities
scipy>=1.11.0

